# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alignment Economy Node API",
    "version": "0.1.0",
    "description": "The Alignment Economy protocol node API. Write requests are signed: the body is a { accountId, timestamp, signature, payload } envelope, where signature is an ML-DSA-65 signature over the canonical payload. Amounts are base-unit integer strings (10^8 per point)."
  },
  "servers": [
    {
      "url": "https://alignmenteconomy.org"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/v1/court/challenges": {
      "post": {
        "summary": "File a challenge (miner only)",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountId",
                  "timestamp",
                  "signature",
                  "payload"
                ],
                "properties": {
                  "accountId": {
                    "type": "string"
                  },
                  "timestamp": {
                    "type": "number",
                    "description": "Unix seconds; must be within 5 minutes of server time"
                  },
                  "signature": {
                    "type": "string",
                    "description": "ML-DSA-65 signature over the canonical payload + timestamp"
                  },
                  "payload": {
                    "type": "object",
                    "properties": {
                      "defendantAccountId": {
                        "type": "string",
                        "minLength": 1
                      },
                      "caseType": {
                        "type": "string",
                        "enum": [
                          "not_human",
                          "duplicate_account"
                        ]
                      },
                      "stakePercent": {
                        "type": "number"
                      },
                      "openingArgument": {
                        "type": "string"
                      },
                      "counterpartAccountId": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "defendantAccountId",
                      "caseType",
                      "stakePercent"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "signedEnvelope": []
          }
        ]
      }
    },
    "/api/v1/court/cases/{id}/arguments": {
      "post": {
        "summary": "Submit a case argument",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountId",
                  "timestamp",
                  "signature",
                  "payload"
                ],
                "properties": {
                  "accountId": {
                    "type": "string"
                  },
                  "timestamp": {
                    "type": "number",
                    "description": "Unix seconds; must be within 5 minutes of server time"
                  },
                  "signature": {
                    "type": "string",
                    "description": "ML-DSA-65 signature over the canonical payload + timestamp"
                  },
                  "payload": {
                    "type": "object",
                    "properties": {
                      "text": {
                        "type": "string",
                        "minLength": 1
                      },
                      "attachmentHash": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "text"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "signedEnvelope": []
          }
        ]
      }
    },
    "/api/v1/court/cases/{id}/vote": {
      "post": {
        "summary": "Cast a juror vote (miner only)",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountId",
                  "timestamp",
                  "signature",
                  "payload"
                ],
                "properties": {
                  "accountId": {
                    "type": "string"
                  },
                  "timestamp": {
                    "type": "number",
                    "description": "Unix seconds; must be within 5 minutes of server time"
                  },
                  "signature": {
                    "type": "string",
                    "description": "ML-DSA-65 signature over the canonical payload + timestamp"
                  },
                  "payload": {
                    "type": "object",
                    "properties": {
                      "vote": {
                        "type": "string",
                        "enum": [
                          "human",
                          "not_human"
                        ]
                      }
                    },
                    "required": [
                      "vote"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "signedEnvelope": []
          }
        ]
      }
    },
...
    "/api/v1/court/cases/{id}": {
      "get": {
        "summary": "Get a court case",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "signedEnvelope": {
        "type": "apiKey",
        "in": "header",
        "name": "signature",
        "description": "Authentication is a signed request envelope, not a header token: send { accountId, timestamp, signature, payload } where signature is the ML-DSA-65 signature over the payload. Documented as apiKey for tooling compatibility."
      }
    }
  }
}